I have divide the code into different layes for more understable
