+++ Interface/IActivity.cs
@@ -1,0 +1,7 @@
+﻿namespace Intermediate_Excercises.Interface
+{
+    internal interface IActivity
+    {
+        public void Execute();
+    }
+}
