+++ Intermediate_Excercises.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35931.197 d17.13
+VisualStudioVersion = 17.13.35931.197
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Intermediate_Excercises", "Intermediate_Excercises.csproj", "{36F1909D-EE02-47BE-A1F3-6B7F678DA432}"
 EndProject
