+++ Models/Post.cs
@@ -1,0 +1,31 @@
+﻿namespace Intermediate_Excercises.Models
+{
+    //Q#2 Simulating how the votes on post of stackoverflow works 
+    public class Post
+    {
+        public string Title { get; set; }
+        public string Description { get; set; }
+        public DateTime Created { get; }
+        public int votecount;
+
+        public Post(string title, string description)
+        {
+            Title = title;
+            Description = description;
+            Created = DateTime.Now;
+            votecount = 0;
+        }
+
+        public int Votecount { get { return votecount; } }
+
+        public void up_Vote()
+        {
+            votecount++;
+        }
+        //
+        public void down_Vote()
+        {
+            votecount--;
+        }
+    }
+}
